requirements.txt
Pillow
matplotlib
tensorflow
numpy
